Unify traces description Change showing device's id to grouping statistics by device's description. Fixes https://github.com/Open-CAS/standalone-linux-io-tracer/issues/166
